Fix UI and DB

+++ database/migrations/2022_03_30_082903_scores_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id()->autoIncrement()->nullable();
             $table->string('tenmh');
             $table->integer('sotc');
-            $table->integer('diem');
+            $table->float('diem');
             $table->string('masv');
             $table->string('lop');
             $table->integer('lanthi');
